% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_all.R
\name{build_all}
\alias{build_all}
\title{Build all pieces}
\usage{
build_all(dir = NULL, clean = FALSE)
}
\arguments{
\item{dir}{Directory in which to create images}

\item{clean}{Delete any existing files in directory? (default = FALSE)}
}
\description{
Build all pieces
}
