% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxious_morning.R, R/arrival_in_montreal.R,
%   R/zzz.R
\name{anxious_morning}
\alias{anxious_morning}
\alias{arrival_in_montreal}
\alias{rosemary}
\title{Generative artwork}
\usage{
anxious_morning(dir = NULL, ...)

arrival_in_montreal(dir = NULL, ...)
}
\arguments{
\item{dir}{Directory to write file to}

\item{...}{Other arguments}
}
\description{
Generative artwork
}
