% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxious_morning.R
\name{rosemary}
\alias{rosemary}
\alias{anxious_morning}
\title{Generative artwork}
\usage{
anxious_morning(dir = NULL, ...)
}
\arguments{
\item{dir}{Directory to write file to}

\item{...}{Other arguments}
}
\description{
Generative artwork
}
