% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxious_morning.R, R/arrival_in_montreal.R,
%   R/art_deco_rocketry.R, R/zzz.R
\name{anxious_morning}
\alias{anxious_morning}
\alias{arrival_in_montreal}
\alias{art_deco_rocketry}
\alias{rosemary}
\title{Generative artwork}
\usage{
anxious_morning(dir = NULL, ...)

arrival_in_montreal(dir = NULL, ...)

art_deco_rocketry(dir = NULL, version = 0, ...)
}
\arguments{
\item{dir}{Directory to write file to}

\item{...}{Other arguments}

\item{version}{Which version to write? (0 means create all)}
}
\description{
Generative artwork
}
