% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxious_morning.R, R/arrival_in_montreal.R,
%   R/art_deco_rocketry.R, R/bisexual_flag.R, R/constellations.R,
%   R/crow_rings.R, R/cubismic.R, R/delaunay_duck.R, R/devil.R,
%   R/eclipse.R, R/electric_princess.R, R/forgotten_worlds.R, R/goodnight.R,
%   R/heart_sketch.R, R/ice_queen.R, R/incantations.R, R/light_blades.R,
%   R/mermaid.R, R/mist_wraith.R, R/molten_heart.R, R/neon_heart.R,
%   R/neon_lights.R, R/orange_tear.R, R/pointillist.R, R/press_play.R,
%   R/rainbow_prisms.R, R/rainbow_tears.R, R/screw_the_roses.R, R/splash.R,
%   R/sponge.R, R/summoning_circles.R, R/the_art_of_escaping_boxes.R,
%   R/thin_swirls.R, R/transgender_flag.R, R/warped_heart.R, R/white_rose.R,
%   R/wisps.R, R/wispy_heart.R, R/zzz.R
\name{anxious_morning}
\alias{anxious_morning}
\alias{arrival_in_montreal}
\alias{art_deco_rocketry}
\alias{bisexual_flag}
\alias{constellations}
\alias{crow_rings}
\alias{cubismic}
\alias{delaunay_duck}
\alias{devil}
\alias{eclipse}
\alias{electric_princess}
\alias{forgotten_worlds}
\alias{goodnight}
\alias{heart_sketch}
\alias{ice_queen}
\alias{incantations}
\alias{light_blades}
\alias{mermaid}
\alias{mist_wraith}
\alias{molten_heart}
\alias{neon_heart}
\alias{neon_lights}
\alias{orange_tear}
\alias{pointillist}
\alias{press_play}
\alias{rainbow_prisms}
\alias{rainbow_tears}
\alias{screw_the_roses}
\alias{splash}
\alias{sponge}
\alias{summoning_circles}
\alias{the_art_of_escaping_boxes}
\alias{thin_swirls}
\alias{transgender_flag}
\alias{warped_heart}
\alias{white_rose}
\alias{wisps}
\alias{wispy_heart}
\alias{rosemary}
\title{Generative artwork}
\usage{
anxious_morning(dir = NULL, ...)

arrival_in_montreal(dir = NULL, ...)

art_deco_rocketry(dir = NULL, version = 0, ...)

bisexual_flag(dir = NULL, ...)

constellations(dir = NULL, ...)

crow_rings(dir = NULL, ...)

cubismic(dir = NULL, version = 0, ...)

delaunay_duck(dir = NULL, ...)

devil(dir = NULL, ...)

eclipse(dir = NULL, ...)

electric_princess(dir = NULL, ...)

forgotten_worlds(dir = NULL, ...)

goodnight(dir = NULL, ...)

heart_sketch(dir = NULL, ...)

ice_queen(dir = NULL, ...)

incantations(dir = NULL, ...)

light_blades(dir = NULL, ...)

mermaid(dir = NULL, ...)

mist_wraith(dir = NULL, ...)

molten_heart(dir = NULL, version = 0, ...)

neon_heart(dir = NULL, ...)

neon_lights(dir = NULL, ...)

orange_tear(dir = NULL, ...)

pointillist(dir = NULL, ...)

press_play(dir = NULL, ...)

rainbow_prisms(dir = NULL, ...)

rainbow_tears(dir = NULL, ...)

screw_the_roses(dir = NULL, ...)

splash(dir = NULL, ...)

sponge(dir = NULL, ...)

summoning_circles(dir = NULL, ...)

the_art_of_escaping_boxes(dir = NULL, ...)

thin_swirls(dir = NULL, ...)

transgender_flag(dir = NULL, ...)

warped_heart(dir = NULL, ...)

white_rose(dir = NULL, ...)

wisps(dir = NULL, ...)

wispy_heart(dir = NULL, version = 0, ...)
}
\arguments{
\item{dir}{Directory to write file to}

\item{...}{Other arguments}

\item{version}{Which version to write? (0 means create all)}
}
\description{
Generative artwork
}
