% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_square_of_balls.R, R/anguish.R,
%   R/anxious_morning.R, R/arrival_in_montreal.R, R/art_032.R,
%   R/art_deco_rocketry.R, R/bisexual_flag.R, R/bloody_valentine.R,
%   R/constellations.R, R/crow_rings.R, R/cubismic.R, R/cylon_raider.R,
%   R/delaunay_duck.R, R/devil.R, R/electric_princess.R, R/forgotten_worlds.R,
%   R/fracture.R, R/freecell.R, R/goodnight.R, R/heart_sketch.R,
%   R/ice_queen.R, R/incantations.R, R/light_blades.R, R/mermaid.R,
%   R/mist_wraith.R, R/molten_heart.R, R/neon_heart.R, R/neon_lights.R,
%   R/nine_hearts.R, R/one_art_please.R, R/orange_tear.R, R/pointillist.R,
%   R/press_play.R, R/purple_heart.R, R/rainbow_prisms.R, R/rainbow_tears.R,
%   R/regret.R, R/screw_the_roses.R, R/sea_snakes.R, R/sponge.R,
%   R/stars_and_the_night_sky.R, R/summoning_circles.R, R/tendrils_of_memory.R,
%   R/the_art_of_escaping_boxes.R, R/thin_swirls.R, R/transgender_flag.R,
%   R/warped_heart.R, R/wisps.R, R/wispy_heart.R, R/zzz.R
\name{a_square_of_balls}
\alias{a_square_of_balls}
\alias{anguish}
\alias{anxious_morning}
\alias{arrival_in_montreal}
\alias{art_032}
\alias{art_deco_rocketry}
\alias{bisexual_flag}
\alias{bloody_valentine}
\alias{constellations}
\alias{crow_rings}
\alias{cubismic}
\alias{cylon_raider}
\alias{delaunay_duck}
\alias{devil}
\alias{electric_princess}
\alias{forgotten_worlds}
\alias{fracture}
\alias{freecell}
\alias{goodnight}
\alias{heart_sketch}
\alias{ice_queen}
\alias{incantations}
\alias{light_blades}
\alias{mermaid}
\alias{mist_wraith}
\alias{molten_heart}
\alias{neon_heart}
\alias{neon_lights}
\alias{nine_hearts}
\alias{one_art_please}
\alias{orange_tear}
\alias{pointillist}
\alias{press_play}
\alias{purple_heart}
\alias{rainbow_prisms}
\alias{rainbow_tears}
\alias{regret}
\alias{screw_the_roses}
\alias{sea_snakes}
\alias{sponge}
\alias{stars_and_the_night_sky}
\alias{summoning_circles}
\alias{tendrils_of_memory}
\alias{the_art_of_escaping_boxes}
\alias{thin_swirls}
\alias{transgender_flag}
\alias{warped_heart}
\alias{wisps}
\alias{wispy_heart}
\alias{rosemary}
\title{Generative artwork}
\usage{
a_square_of_balls(dir = NULL, ...)

anguish(dir = NULL, ...)

anxious_morning(dir = NULL, ...)

arrival_in_montreal(dir = NULL, ...)

art_032(dir = NULL, ...)

art_deco_rocketry(dir = NULL, version = 0, ...)

bisexual_flag(dir = NULL, ...)

bloody_valentine(dir = NULL, ...)

constellations(dir = NULL, ...)

crow_rings(dir = NULL, ...)

cubismic(dir = NULL, version = 0, ...)

cylon_raider(dir = NULL, ...)

delaunay_duck(dir = NULL, ...)

devil(dir = NULL, ...)

electric_princess(dir = NULL, ...)

forgotten_worlds(dir = NULL, ...)

fracture(dir = NULL, ...)

freecell(dir = NULL, ...)

goodnight(dir = NULL, ...)

heart_sketch(dir = NULL, ...)

ice_queen(dir = NULL, ...)

incantations(dir = NULL, ...)

light_blades(dir = NULL, ...)

mermaid(dir = NULL, ...)

mist_wraith(dir = NULL, ...)

molten_heart(dir = NULL, version = 0, ...)

neon_heart(dir = NULL, ...)

neon_lights(dir = NULL, ...)

nine_hearts(dir = NULL, ...)

one_art_please(dir = NULL, ...)

orange_tear(dir = NULL, ...)

pointillist(dir = NULL, ...)

press_play(dir = NULL, ...)

purple_heart(dir = NULL, ...)

rainbow_prisms(dir = NULL, ...)

rainbow_tears(dir = NULL, ...)

regret(dir = NULL, ...)

screw_the_roses(dir = NULL, ...)

sea_snakes(dir = NULL, ...)

sponge(dir = NULL, ...)

stars_and_the_night_sky(dir = NULL, ...)

summoning_circles(dir = NULL, ...)

tendrils_of_memory(dir = NULL, version = 0, ...)

the_art_of_escaping_boxes(dir = NULL, ...)

thin_swirls(dir = NULL, ...)

transgender_flag(dir = NULL, ...)

warped_heart(dir = NULL, ...)

wisps(dir = NULL, ...)

wispy_heart(dir = NULL, version = 0, ...)
}
\arguments{
\item{dir}{Directory to write file to}

\item{...}{Other arguments}

\item{version}{Which version to write? (0 means create all)}
}
\description{
Generative artwork
}
