% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxious_morning.R, R/arrival_in_montreal.R,
%   R/art_deco_rocketry.R, R/bisexual_flag.R, R/constellations.R,
%   R/crow_rings.R, R/cubismic.R, R/delaunay_duck.R, R/devil.R, R/eclipse.R,
%   R/electric_princess.R, R/zzz.R
\name{anxious_morning}
\alias{anxious_morning}
\alias{arrival_in_montreal}
\alias{art_deco_rocketry}
\alias{bisexual_flag}
\alias{constellations}
\alias{crow_rings}
\alias{cubismic}
\alias{delaunay_duck}
\alias{devil}
\alias{eclipse}
\alias{electric_princess}
\alias{rosemary}
\title{Generative artwork}
\usage{
anxious_morning(dir = NULL, ...)

arrival_in_montreal(dir = NULL, ...)

art_deco_rocketry(dir = NULL, version = 0, ...)

bisexual_flag(dir = NULL, ...)

constellations(dir = NULL, ...)

crow_rings(dir = NULL, ...)

cubismic(dir = NULL, version = 0, ...)

delaunay_duck(dir = NULL, ...)

devil(dir = NULL, ...)

eclipse(dir = NULL, ...)

electric_princess(dir = NULL, ...)
}
\arguments{
\item{dir}{Directory to write file to}

\item{...}{Other arguments}

\item{version}{Which version to write? (0 means create all)}
}
\description{
Generative artwork
}
